Use full patch version for go directives in go.mod
It turns out that since Go 1.21, in which Go started using fully
qualified versions and not omitting zeros (i.e. "1.21.0" instead of just
"1.21"), just stating the minor version without the patch version is
actually considered a development version of Go. Who would have thought?
The correct way to use the Go directive is to set it to "go 1.23.0".

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>